+++ packages/common/src/project/readers/ipfs-reader.ts
@@ -17,7 +17,10 @@ export class IPFSReader implements Reader {
   private ipfs: IPFSHTTPClient;
   private cache: Record<string, Promise<string>> = {};
 
-  constructor(private readonly cid: string, gateway?: string) {
+  constructor(
+    private readonly cid: string,
+    gateway?: string
+  ) {
     if (!CIDv0.test(cid) && !CIDv1.test(cid)) {
       throw new Error('IPFS project path CID is not valid');
     }
@@ -28,19 +31,17 @@ export class IPFSReader implements Reader {
     return undefined;
   }
 
-  async getPkg(): Promise<IPackageJson | undefined> {
-    return Promise.resolve(undefined);
+  // eslint-disable-next-line @typescript-eslint/require-await
+  async getPkg(): Promise<IPackageJson> {
+    throw new Error("IPFS Reader doesn't support reading package.json");
   }
 
-  async getProjectSchema(): Promise<unknown | undefined> {
+  async getProjectSchema(): Promise<unknown> {
     const projectYaml = await this.getFile(this.cid);
-    if (projectYaml === undefined) {
-      throw new Error(`Failed to fetch project from IPFS: ${this.cid}`);
-    }
     return yaml.load(projectYaml);
   }
 
-  async getFile(fileName: string): Promise<string | undefined> {
+  async getFile(fileName: string): Promise<string> {
     try {
       if (this.cache[fileName] === undefined) {
         const resolvedFileName = fileName.replace('ipfs://', '');
@@ -48,8 +49,8 @@ export class IPFSReader implements Reader {
       }
       return await this.cache[fileName];
     } catch (e) {
-      console.error(`Failed to fetch file from IPFS: ${fileName}`, e);
-      return undefined;
+      console.error(`Failed to fetch project from IPFS: ${fileName}`, e);
+      throw new Error(`Failed to fetch project from IPFS: ${fileName}`, {cause: e});
     }
   }