+++ packages/node-core/CHANGELOG.md
@@ -6,11 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [10.9.0] - 2024-06-21
 ### Fixed
 - Store service index check failing with snake_case fields (#2461)
 
-## Changed
-- Error message for finding available port 
+### Changed
+- Error message for finding available port
 
 ## [10.6.0] - 2024-06-18
 ### Fixed
@@ -738,7 +739,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Move blockchain agnostic code from `node` to `node-core` package. (#1222)
 
-[Unreleased]: https://github.com/subquery/subql/compare/node-core/10.6.0...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/node-core/10.9.0...HEAD
+[10.9.0]: https://github.com/subquery/subql/compare/node-core/10.6.0...node-core/10.9.0
 [10.6.0]: https://github.com/subquery/subql/compare/node-core/10.5.1...node-core/10.6.0
 [10.5.1]: https://github.com/subquery/subql/compare/node-core/10.5.0...node-core/10.5.1
 [10.5.0]: https://github.com/subquery/subql/compare/node-core/10.4.1...node-core/10.5.0